fix(payload): perform gas metric subtraction in f64 space
Prevents panic from integer subtraction overflow in debug/test builds
by casting to f64 before subtracting. Negative values surface upstream
bugs on the dashboard instead of silently clamping to zero.

Amp-Thread-ID: https://ampcode.com/threads/T-019d047d-2aad-775b-8c54-b3d9e646beee

+++ crates/payload/builder/src/lib.rs
@@ -630,13 +630,13 @@ where
             .set(non_payment_gas_used as f64);
         self.metrics
             .payment_gas_used_last
-            .set((cumulative_gas_used - non_payment_gas_used) as f64);
+            .set(cumulative_gas_used as f64 - non_payment_gas_used as f64);
         self.metrics
             .general_gas_limit_last
             .set(general_gas_limit as f64);
         self.metrics
             .payment_gas_limit_last
-            .set((non_shared_gas_limit - general_gas_limit) as f64);
+            .set(non_shared_gas_limit as f64 - general_gas_limit as f64);
         self.metrics
             .shared_gas_limit_last
             .set(shared_gas_limit as f64);
